requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pgvector==0.2.5

dramatiq==1.17.0
redis==5.0.8

Pillow==10.4.0
numpy==1.26.4
torch==2.3.1
open-clip-torch==2.24.0

# scraping
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2
